modify the vote from double to int64 optimization for vote delegate

+++ consensus/rpos/include/rpos/CommitteeInfo.h
@@ -5,7 +5,7 @@ namespace ultrainio {
     struct CommitteeInfo {
         std::string accountName;
         std::string pk;
-        double stakesCount;
+        int64_t stakesCount;
     };
 }
